+++ CHANGELOG.rst
@@ -24,6 +24,8 @@ Release 0.116.0
 - Fix token authentication according to DRF 3.3.
 - Replace IPAddressField serializer with DRF serializer.
 - Add get_extra_kwargs method to AugmentedSerializerMixin.
+- Migrate to DRF 3.5.3 and django_filter 0.15.3.
+- Drop DjangoMappingFilterBackend and UUIDFilter.
 
 Release 0.115.0
 ---------------